requirements.txt
beautifulsoup4==4.11.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
contextvars==2.4
cymem==2.0.6
dataclasses==0.8
distro==1.7.0
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.0/en_core_web_trf-3.4.0-py3-none-any.whl
filelock==3.4.1
Flask==2.0.3
Flask-Cors==3.0.10
huggingface-hub==0.4.0
idna==3.3
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.7
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pathy==0.6.2
preshed==3.0.6
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.25
requests==2.27.1
sacremoses==0.0.53
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-alignments==0.8.5
spacy-legacy==3.0.9
spacy-loggers==1.0.3
spacy-transformers==1.1.7
srsly==2.4.4
tabula-py==2.3.0
thinc==8.1.0
tokenizers==0.12.1
torch==1.10.2
tqdm==4.64.0
transformers==4.18.0
typer==0.4.2
typing_extensions==4.1.1
urllib3==1.26.11
wasabi==0.10.1
Werkzeug==2.0.3
zipp==3.6.0
